pyproject.toml
[build-system]
requires = [
    "setuptools >= 61.0.0",
    "setuptools_scm[toml] >= 7",
    "wheel >= 0.29.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "scotty-beam-tracing"
description = "Beam tracing for tokamak DBS diagnostics"
requires-python = ">3.7"
readme = "README.md"
authors = [{name = "Valerian Hall-Chen"}]
urls = {project = "https://github.com/valerian-chen/Scotty"}
dependencies = [
    "numpy~=1.20",
    "scipy~=1.7",
    "matplotlib~=3.3",
    "freeqdsk==0.1.0",
    "xarray>=2023.01.0",
    "xarray-datatree",
    "h5netcdf",
]
dynamic = ["version"]

[project.optional-dependencies]
tests = [
    "pytest",
]
docs = [
    "sphinx >= 4.0",
    "sphinx-book-theme ~= 0.3.3",
]

[tool.setuptools]
packages = ["scotty"]

[tool.setuptools.dynamic]
version = { attr = "setuptools_scm.get_version" }

[tool.setuptools_scm]
write_to = "scotty/_version.py"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore::UserWarning:h5netcdf",
]

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"
ignore_missing_imports = "True"

[tool.ruff]
line-length = 88
ignore = ["E501"]
